Updating the paths for dependencies within the separate files
