+++ src/Schedule.C
@@ -14,7 +14,7 @@
 #include "ScheduleItem.h"
 #include "CoolingTime.h"
 
-#include "Calc/topSchedule.h"
+#include "topSchedule.h"
 
 /***************************
  ********* Service *********
