+++ src/ScheduleItem.C
@@ -15,7 +15,7 @@
 #include "History.h"
 #include "Schedule.h"
 #include "Flux.h"
-#include "Calc/calcSchedule.h"
+#include "calcSchedule.h"
 
 /***************************
  ********* Service *********
