+++ src/topScheduleT.C
@@ -10,7 +10,7 @@
 #include "topScheduleT.h"
 #include "topSchedule.h"
 
-#include "Chains/NuclearData.h"
+#include "NuclearData.h"
 
 /****************************
  ********* Service **********
